discord_bot/requirements.txt
discord.py[voice]
httpx
python-dotenv